FIXUP: put constant and white noise in kernel.py and fix docs
