+++ tests/test_gp.py
@@ -30,7 +30,7 @@ COV_FUNCS = {
 # NOTE: the mean and covariance functions used here
 # must be present in `MEAN_FUNCS` and `COV_FUNCS` resp.
 GP_MODELS = {
-    "LatentGP": {"mean_fn": ["Zero", "Zero"], "cov_fn": ["ExpQuad", "Constant"]},
+    "LatentGP": {"mean_fn": ["Zero", "Zero", "Zero"], "cov_fn": ["ExpQuad", "Constant", "WhiteNoise"]},
 }
